refactor(dom): extract the shared DOM translation into @dunky.dev/state-machine-dom
The aria-* attribute projection and the payload adapters were byte-identical
in the React and Solid normalizers, and had already drifted once (the
preventDefault bind fix landed in one but not the other). They now live once
in packages/dom; each target keeps only what genuinely differs — its handler
prop names (onChange/onDoubleClick vs onInput/onDblClick), the focusable →
tabindex casing, and its value serialization (React passes ARIA booleans
through, Solid stringifies them).

Payload construction is pinned once in the dom package's tests; the target
suites keep a wiring proof each (normalize wraps the handler with its
adapter) plus the vocabulary accounting. No consumer-facing API change.
