+++ sandbox/solid/vite.config.ts
@@ -16,6 +16,7 @@ export default defineConfig({
       '@dunky.dev/solid-state-machine': resolve(__dirname, '../../packages/solid/src'),
       '@dunky.dev/state-machine-utils': resolve(__dirname, '../../packages/shared/utils/src'),
       '@dunky.dev/state-machine-bindings': resolve(__dirname, '../../packages/shared/bindings/src'),
+      '@dunky.dev/state-machine-dom': resolve(__dirname, '../../packages/dom/src'),
       '@sandbox/cmdk-core': resolve(__dirname, '../shared/src'),
     },
   },
