+++ tsdown.config.ts
@@ -11,6 +11,7 @@ export default defineConfig({
   // choice. Keep in sync with the publish set in .changeset/config.json.
   workspace: [
     'packages/core',
+    'packages/dom',
     'packages/react',
     'packages/solid',
     'packages/native',
